* stage: (1061 commits)
  [R2-1803]: Move Icons ui-kit/icons
  Fix setApartmentPauseService
  Update
  [R2-1929]: Update types
  [R2-1803]: Rewrite ResourceIconLookup to IconLookup
  [R2-1807]: Update indexes
  [R2-1856]: Review refactor
  [R2-1803]: Review refactor
  [R2-1803]: Review refactor
  [R2-1807]: Return <div />
  [R2-1807]: Review refactor
  Update README.md
  Back devsettings
  [R2-1907]: Complete
  Fix
  Fix
  [R2-1907]: Remove console.log
  Fixes
  [R2-1907]: Refactor service
  [R2-1910]: Add SetNextStageDeadline service
  ...

+++ src/01/axios.ts
@@ -1,21 +1,13 @@
-import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
+import axiosHttpClient from 'axios';
+import { createEvent, createStore } from 'effector';
 
-const devUrl = 'https://transparent-staging.herokuapp.com/api';
+const devUrl = 'https://management.staging.transparent-technology.ru/api/';
 const baseURL = process.env.REACT_APP_API_URL || devUrl;
 
-axios.defaults.baseURL = baseURL;
+axiosHttpClient.defaults.baseURL = baseURL;
 
-let cancel;
-axios.interceptors.request.use((req) => {
-  if (req.baseURL === 'http://84.201.132.164:8080/api') {
-    delete req.headers.Authorization;
-  } else {
-    req.headers.Authorization = `Bearer ${takeFromLocStor('token')}`;
-  }
-
-  req.cancelToken = new axios.CancelToken((e) => {
-    cancel = e;
-  });
+axiosHttpClient.interceptors.request.use((req) => {
+  req.headers.Authorization = `Bearer ${takeFromLocStor('token')}`;
 
   if (req.url && checkUrl('refresh', req.url)) {
     req.data = {
@@ -26,72 +18,84 @@ axios.interceptors.request.use((req) => {
   return req;
 });
 
-axios.interceptors.response.use(
+axiosHttpClient.interceptors.response.use(
   ({ data, config }) => {
     const { url } = config;
 
-    // const url = config?.url;
     if (url && checkUrl('(login|refresh)', url)) {
       const { token, refreshToken, roles, permissions } = data.successResponse;
-      saveToLocStor('token', token);
-      saveToLocStor('refreshToken', refreshToken);
-      saveToLocStor('permissions', permissions);
-      checkUrl('login', url) && saveToLocStor('roles', roles);
+
+      saveToLocalStorage('token', token);
+      saveToLocalStorage('refreshToken', refreshToken);
+      saveToLocalStorage('permissions', permissions);
+      checkUrl('login', url) && saveToLocalStorage('roles', roles);
     }
 
     if (url && checkUrl('(users/current)', url)) {
       const user = data.successResponse;
-      saveToLocStor('user', user);
+      saveToLocalStorage('user', user);
     }
-    const res = data.successResponse ?? data ?? {};
-    // return { ...res, url };
-    return res;
 
-    // return data
+    const res = data.successResponse ?? data ?? {};
+
+    return res;
   },
   (error) => {
     const status = error?.response?.status;
     if (status === 401 && !checkUrl('login', error.config.url)) {
       const { config } = error;
-      return new Promise((resolve, reject) => {
-        axios.post('/auth/refreshToken').then(
-          () => resolve(axios(config)),
-          () => {
-            localStorage.clear();
-            window.location.replace('/login');
-          }
-        );
+
+      return new Promise((resolve) => {
+        if (!$isRefreshRunning.getState()) {
+          setIsRefreshRunning(true);
+          axiosHttpClient.post('/auth/refreshToken').then(
+            () => {
+              setIsRefreshRunning(false);
+
+              return resolve(axiosHttpClient(config));
+            },
+            () => {
+              localStorage.clear();
+              window.location.replace('/login');
+            }
+          );
+        } else {
+          const subscription = $isRefreshRunning.watch((isRefreshStop) => {
+            if (!isRefreshStop) {
+              resolve(axiosHttpClient(config));
+              subscription.unsubscribe();
+            }
+          });
+        }
       });
     }
 
-    //   if (status === 403 ) {
-    //       window.location.replace('/access-denied/');
-    //   }
-    //   if (status === 404 ) {
-    //       window.location.replace('/error/');
-    //   }
     return Promise.reject(error);
   }
 );
 
-// get<T = any, R = AxiosResponse<T>>(url: string, config?: AxiosRequestConfig): Promise<R>;
-
-// export type CustomGetResponse<T> = Promise<AxiosResponse<T>>;
-
-// utils
-function saveToLocStor(name: string, data: string) {
+function saveToLocalStorage(name: string, data: string) {
   localStorage.setItem(name, JSON.stringify(data));
 }
 
 function takeFromLocStor(name: string) {
-  const userName = localStorage.getItem(name);
-  if (!userName) return;
-  return JSON.parse(userName);
+  const value = localStorage.getItem(name);
+
+  if (!value) return;
+
+  return JSON.parse(value);
 }
 
 function checkUrl(str: string, url: string) {
   return new RegExp(str, 'gi').test(url);
 }
 
-export { cancel };
-export default axios;
+export const $isRefreshRunning = createStore(false);
+
+export const setIsRefreshRunning = createEvent<boolean>();
+
+$isRefreshRunning.on(setIsRefreshRunning, (_, value) => value);
+
+export default axiosHttpClient;
+
+export const axios = axiosHttpClient;
